Move all generator related functions to utilities. Update popup styles
